refactor: 카드 템플릿 메인 컨텐츠 생성 로직 분리
