+++ src/templates/index.ts
@@ -1,4 +1,4 @@
 import DefaultHeader from './DefaultHeader/DefaultHeader';
-import { ProductCard } from './ProductCard';
+import { ProductCard } from './ProductCard/ProductCard';
 
 export { DefaultHeader, ProductCard };
